requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0
aiohttp>=3.9
matplotlib>=3.8
mplfinance>=0.12.10b0
pandas>=2.1
numpy>=1.26
requests>=2.31
pillow>=10.0
yarl>=1.9
typing-extensions>=4.9
aiofiles>=23.1